client-api: Move static profile fields to submodule
Allowing to apply `allow(clippy::exhaustive_structs)` to the whole
module instead of to the individual structs.
